requirements.txt
timm>=0.9.16
torch>=2.2.0
torchvision>=0.17.0
numpy>=1.24.0
pandas>=2.0.0
pillow>=10.0.0
scikit-learn>=1.3.0
PyYAML>=6.0
tqdm>=4.66.0
opencv-python-headless>=4.8.0.76
streamlit>=1.28.0
